code cleanup for sparse/linear kernels

+++ src/kernel/PolyKernel.h
@@ -20,9 +20,7 @@ class CPolyKernel: public CSimpleKernel<DREAL>
  public:
 
   CPolyKernel(CRealFeatures* l, CRealFeatures* r, INT size, INT d, bool inhom, bool use_norm);
-
-  CPolyKernel(LONG size, INT degree, bool inhomogene=true, bool use_normalization=true);
-
+  CPolyKernel(INT size, INT degree, bool inhomogene=true, bool use_normalization=true);
   ~CPolyKernel();
   
   virtual bool init(CFeatures* l, CFeatures* r);
